Remove useless registry key `addedtopath'
Its only usefulness is indicating whether use has added winchecksum to
Path. However, this key will be deleted along with parent key when
uninstalling, but uninstallation won't remove winchecksum from Path. As
a result, it shows wrong information after uninstalling.

If we remove winchecksum from path when installing, another issue would
appear. When installing, a uninstallation is execute silently to ensure
that reinstalling is clean, and user would have to add winchecksum to
path again after a reinstallation, which should by no means happen
because adding to Path and installing are irrelevant.